#  Improve the iterate seek to O(logN)