Hi, In our scenario, we frequently use the iterate, and the `Less` compare is expenses when key is large.   This PR mainly improve the iterate location. uses binary search to quickly locates the start item.
before:
```
BenchmarkSeek-4          3000000               452 ns/op
BenchmarkSeek-4          3000000               513 ns/op
BenchmarkSeek-4          3000000               454 ns/op
BenchmarkSeek-4          3000000               450 ns/op
```
after:
```
BenchmarkSeek-4         10000000               225 ns/op
BenchmarkSeek-4         10000000               222 ns/op
BenchmarkSeek-4         10000000               223 ns/op
BenchmarkSeek-4         10000000               224 ns/op
```